requirement.txt
streamlit
openai
requests
python-dotenv
pandas
python-dateutil